requirements.txt
gunicorn
Flask==3.0.3
supabase==2.6.0
python-dotenv==1.0.1
